feat(workflow): migrate if-else, loop and human-confirmation to the catalog
The branch-node batch (11 of 38 types migrated) - the first manifests with
real connection.branches() implementations, the contract the parity suite's
handle assertions and the future connect_nodes tool read from:

- if-else derives one branch per case plus the reserved 'false' ELSE,
  through branchNameCorrect (relocated into the manifest file; the canvas's
  calculateTargetBranches keeps its own copy of the derivation until
  text-classifier/http/crud migrate and both converge on the manifest).
- human-confirmation declares its three-way approved/denied/timeout routing.
- loop declares source + loop-start (loop-back is a TARGET handle and
  documented as such), and its agent usage teaches the three-edge shape,
  parentId containment, and node-scoped item refs only.

Also in this batch: TargetBranch/BranchType relocate to catalog/node-base
(web types/core.ts re-exports); LOOP_CONSTANTS/LOOP_HANDLES move into the
loop manifest file while the builder-only LOOP_VARIABLES vocabulary stays
web-side with a drift warning (the engine never writes loop.*); the engine's
flow-nodes/loop.ts LoopNodeConfig shadow becomes a Pick of the catalog data;
and two more schema-level completeness rules moved to validators where the
defaults-parse discipline flagged them (loop's itemsSource min(1),
human-confirmation's assignee .refine - both defaults legitimately ship
unconfigured). The deprecated consumer-less ifElseSchema died in the move.

Zero diff under any node.tsx / panel.tsx / trace-renderer.tsx.
