+++ apps/web/src/components/workflow/nodes/core/if-else/types.ts
@@ -1,11 +1,23 @@
 // apps/web/src/components/workflow/nodes/core/if-else/types.ts
 
 import type { Operator } from '@auxx/lib/conditions/client'
+import type {
+  CatalogIfElseNodeData,
+  NodeCondition as CatalogNodeCondition,
+} from '@auxx/lib/workflow-engine/client'
 import type { Node as FlowNode } from '@xyflow/react'
 import type { TargetBranch } from '~/components/workflow/types'
-import type { BaseNodeData, SpecificNode } from '~/components/workflow/types/node-base'
+import type { SpecificNode } from '~/components/workflow/types/node-base'
+import type { NodeType } from '~/components/workflow/types/node-types'
 import { BaseType, type UnifiedVariable } from '~/components/workflow/types/variable-types'
 import type { TiptapJSON } from '~/components/workflow/ui/input-editor'
+
+// The data half moved to the node catalog (node-catalog Phase 1 —
+// `@auxx/lib/workflow-engine/catalog/nodes/if-else`). This file keeps the
+// web-only narrowings: `TiptapJSON` in condition values (lib types them as
+// `Record<string, any>`, which Tiptap docs satisfy structurally) and the
+// NodeType narrowing on the node data.
+export type { NodeCondition } from '@auxx/lib/workflow-engine/client'
 
 export type Node = FlowNode
 export type ValueSelector = string[]
@@ -20,10 +32,22 @@ export enum LogicalOperator {
 }
 
 // Extended condition interface for if-else nodes
-export interface IfElseCondition extends Omit<NodeCondition, 'value'> {
+export interface IfElseCondition extends Omit<CatalogNodeCondition, 'value' | 'varType'> {
   file_var?: any
   conditions?: IfElseCondition[]
   value?: string | number | boolean | string[] | TiptapJSON
+  /** Declared value type — drives which value editor is rendered */
+  varType?: BaseType
+}
+
+/**
+ * Case definition for if-else nodes
+ */
+export interface NodeCase {
+  id: string
+  case_id: string
+  logical_operator: 'and' | 'or'
+  conditions: IfElseCondition[]
 }
 
 // Re-export from store types for backward compatibility
@@ -32,11 +56,8 @@ export type IfElseCase = NodeCase
 /**
  * Node data for if-else nodes (flattened structure)
  */
-export interface IfElseNodeData extends BaseNodeData {
-  // Base fields
-  title: string
-  desc?: string
-  // If-else specific fields
+export interface IfElseNodeData extends CatalogIfElseNodeData {
+  type: NodeType
   cases: NodeCase[]
   _targetBranches?: TargetBranch[]
 }
@@ -52,36 +73,6 @@ export type IfElseNode = SpecificNode<'if-else', IfElseNodeData>
 export interface IfElseExecutionResult {
   outputs: { matched_case: string | null; branch: string }
   [key: string]: any
-}
-
-/**
- * Condition for if-else nodes
- * Uses Operator from lib for type safety
- */
-export interface NodeCondition {
-  id: string
-  /** Variable this condition reads. Empty until the user picks one. */
-  variableId?: string
-  comparison_operator?: Operator
-  value?: string | number | boolean | any[] | Record<string, any>
-  /** Whether the right-hand value is a literal rather than a variable reference */
-  isConstant?: boolean
-  /** Sub-key inside a structured variable (e.g. an address part) */
-  key?: string
-  /** Declared value type — drives which value editor is rendered */
-  varType?: BaseType
-  /** How this condition joins the previous one inside its case */
-  logical_operator?: 'and' | 'or'
-}
-
-/**
- * Case definition for if-else nodes
- */
-export interface NodeCase {
-  id: string
-  case_id: string
-  logical_operator: 'and' | 'or'
-  conditions: IfElseCondition[]
 }
 
 // Re-export for convenience
