+++ apps/web/src/components/workflow/nodes/core/loop/constants.ts
@@ -1,11 +1,19 @@
 // apps/web/src/components/workflow/nodes/core/loop/constants.ts
 
-export const LOOP_CONSTANTS = {
-  DEFAULT_MAX_ITERATIONS: 100,
-  ABSOLUTE_MAX_ITERATIONS: 1000,
-  DEFAULT_ITERATOR_NAME: 'item',
-} as const
+// LOOP_CONSTANTS and LOOP_HANDLES moved to the node catalog with loop's data
+// half (node-catalog Phase 1 — `@auxx/lib/workflow-engine/catalog/nodes/loop`);
+// re-exported here so no web import churns.
+export { LOOP_CONSTANTS, LOOP_HANDLES } from '@auxx/lib/workflow-engine/client'
 
+/**
+ * Builder-only loop variable vocabulary.
+ *
+ * ⚠ Known drift (re-verification 2026-08-12): the ENGINE writes the bare
+ * `index` and node-scoped `<loopNodeId>.item`/`.index`/… keys — nothing in the
+ * engine writes these `loop.*` names (only the single-node-run hook seeds
+ * them). Deliberately NOT moved to the catalog until that drift is resolved;
+ * agent-facing guidance teaches node-scoped refs only.
+ */
 export const LOOP_VARIABLES = {
   INDEX: 'loop.index',
   COUNT: 'loop.count',
@@ -14,9 +22,4 @@ export const LOOP_VARIABLES = {
   IS_FIRST: 'loop.isFirst',
   IS_LAST: 'loop.isLast',
   RESULTS: 'loop.results',
-} as const
-
-export const LOOP_HANDLES = {
-  LOOP_START: 'loop-start', // Source handle that connects to first node in loop body
-  LOOP_BACK: 'loop-back', // Target handle where nodes inside loop connect to restart iteration
 } as const
