+++ apps/web/src/components/workflow/parity/catalog-coverage.test.ts
@@ -107,6 +107,6 @@ describe('defineFromManifest', () => {
 
   it('resolves migrated types and stays undefined-safe for unmigrated ones', () => {
     expect(getManifest('wait')?.displayName).toBe('Wait')
-    expect(getManifest('loop')).toBeUndefined()
+    expect(getManifest('code')).toBeUndefined()
   })
 })
